feat(adk): modify internal package
