+++ adk/common/model/openai.go
