+++ adk/common/prints/util.go
