+++ adk/intro/transfer/subagents/chatmodel.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package internal
+package subagents
 
 import (
 	"context"
@@ -26,7 +26,7 @@ import (
 	"github.com/cloudwego/eino/components/tool/utils"
 	"github.com/cloudwego/eino/compose"
 
-	"github.com/cloudwego/eino-examples/adk/internal/model"
+	"github.com/cloudwego/eino-examples/adk/common/model"
 )
 
 type GetWeatherInput struct {
