+++ adk/intro/workflow/sequential/sequential.go
@@ -23,8 +23,8 @@ import (
 
 	"github.com/cloudwego/eino/adk"
 
-	"github.com/cloudwego/eino-examples/adk/internal/prints"
-	"github.com/cloudwego/eino-examples/adk/intro/workflow/sequential/internal"
+	"github.com/cloudwego/eino-examples/adk/common/prints"
+	"github.com/cloudwego/eino-examples/adk/intro/workflow/sequential/subagents"
 )
 
 func main() {
@@ -33,7 +33,7 @@ func main() {
 	a, err := adk.NewSequentialAgent(ctx, &adk.SequentialAgentConfig{
 		Name:        "ResearchAgent",
 		Description: "A sequential workflow for planning and writing a research report.",
-		SubAgents:   []adk.Agent{internal.NewPlanAgent(), internal.NewWriterAgent()},
+		SubAgents:   []adk.Agent{subagents.NewPlanAgent(), subagents.NewWriterAgent()},
 	})
 	if err != nil {
 		log.Fatal(err)
